move compareNodes to knativeItem
