+++ src/servingTree/servingDataProvider.ts
@@ -18,12 +18,13 @@ import * as vscode from 'vscode';
 // import * as validator from 'validator';
 import * as path from 'path';
 import * as yaml from 'yaml';
-import { ServingTreeItem, compareNodes } from './servingTreeItem';
+import { ServingTreeItem } from './servingTreeItem';
 import { Execute, loadItems } from '../cli/execute';
 import { CliExitData } from '../cli/cmdCli';
 import { KnAPI } from '../cli/kn-api';
 import { KubectlAPI } from '../cli/kubectl-api';
 import { ServingContextType } from '../cli/config';
+import { compareNodes } from '../knative/knativeItem';
 import { Service, CreateService, UpdateService } from '../knative/service';
 import { Revision, Items, Traffic } from '../knative/revision';
 import { KnativeServices } from '../knative/knativeServices';
